fix(utils): centralize git calls and panic on non-zero exit codes
