+++ crates/rari-doc/src/pages/types/spa_homepage.rs
@@ -1,14 +1,11 @@
-use std::ffi::OsStr;
 use std::path::Path;
-use std::process::{Command, Output};
 use std::sync::LazyLock;
 
 use chrono::{DateTime, Utc};
 use rari_types::globals::{content_root, content_translated_root};
 use rari_types::locale::Locale;
-use rari_utils::concat_strs;
+use rari_utils::{concat_strs, git::exec_git};
 use regex::Regex;
-use thiserror::Error;
 
 use crate::cached_readers::{blog_files, contributor_spotlight_files};
 use crate::error::DocError;
@@ -19,14 +16,6 @@ use crate::pages::json::{
     HomePageRecentContribution, NameUrl, Parent,
 };
 use crate::pages::page::{Page, PageLike};
-
-#[derive(Debug, Error)]
-enum GitError {
-    #[error("failed to run git: {0}")]
-    Io(#[from] std::io::Error),
-    #[error("git command exited with non-zero code {exit_code}: {stderr}")]
-    CommandFailed { exit_code: String, stderr: String },
-}
 
 pub fn latest_news() -> Result<Vec<HomePageLatestNewsItem>, DocError> {
     Ok(blog_files()
@@ -106,31 +95,7 @@ fn recent_contributions_from_git(
     path: &Path,
     repo: &str,
 ) -> Result<Vec<HomePageRecentContribution>, DocError> {
-    fn exec_git(args: &[impl AsRef<OsStr>], root: impl AsRef<Path>) -> Result<Output, GitError> {
-        let output = Command::new("git")
-            .args(args)
-            .current_dir(root)
-            .output()?;
-        
-        if !output.status.success() {
-            return Err(GitError::CommandFailed {
-                exit_code: output
-                    .status
-                    .code()
-                    .map(|c| c.to_string())
-                    .unwrap_or_else(|| "None".to_string()),
-                stderr: String::from_utf8_lossy(&output.stderr).into_owned(),
-            });
-        }
-
-        Ok(output)
-    }
-
-    let output = exec_git(
-        &["rev-parse", "--show-toplevel"],
-        path,
-    )
-    .expect("failed to execute git rev-parse");
+    let output = exec_git(&["rev-parse", "--show-toplevel"], path);
 
     let repo_root_raw = String::from_utf8_lossy(&output.stdout);
     let repo_root = repo_root_raw.trim();
@@ -144,8 +109,7 @@ fn recent_contributions_from_git(
             "-z",
         ],
         repo_root,
-    )
-    .expect("failed to execute git log");
+    );
 
     let output_str = String::from_utf8_lossy(&output.stdout);
     Ok(output_str
