+++ crates/rari-utils/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod concat;
 pub mod error;
+pub mod git;
 pub mod io;
